Move type c-api functions
